pyproject.toml
[project]
name = "sipin-sip-state-updater"
version = "0.0.1"
description = "Update the state of the ingest of a SIP package"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "viaa-chassis==0.2.1",
    "pulsar-client>=3.8.0,<4.0.0",
    "psycopg>=3.2.9,<4.0.0",
    "psycopg-pool>=3.2.6,<4.0.0",
    "meemoo-cloudevents==0.1.0-rc.4",
    "mediahaven",

]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--rootdir=./app/ -v"
testpaths = ["tests"]
